be: consolidate weather functionality
* move weatherService from noxd/services to noxd/controllers
* add db model for weather entries
